requirements.txt
flask>=3.0,<4.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-wtf>=1.2
flask-migrate>=4.0
flask-limiter>=3.5
sqlalchemy>=2.0
pymysql>=1.1
cryptography>=42.0
python-dotenv>=1.0
twilio>=9.0
openai>=1.30
# googlemaps — optional, install manually if you have a Google Maps API key:
#   pip install "googlemaps>=4.10,<5"
# Without it, server-side geocoding falls back to the browser Autocomplete only.
firebase-admin>=6.4
cloudinary>=1.40
pillow>=10.0
gunicorn>=21.2
itsdangerous>=2.1
werkzeug>=3.0
requests>=2.31
pytest>=8.0
pytest-flask>=1.3
